backend/requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
bcrypt==5.0.0
beautifulsoup4==4.12.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
deepface==0.0.98
Deprecated==1.3.1
deprecation==2.1.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.0
filelock==3.20.3
fire==0.7.1
flatbuffers==25.12.19
fsspec==2026.1.0
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
limits==5.8.0
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
namex==0.1.0
numpy==2.4.2
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
pillow==10.4.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
realtime==2.27.2
requests==2.32.0
retina-face==0.0.17
rich==14.3.2
ruff==0.15.1
setuptools==80.10.2
six==1.17.0
slowapi==0.1.9
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==0.38.6
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.0
vulture==2.14
websockets==15.0.1
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.1.0
yarl==1.22.0
